Switch from Python DevContainer feature to OS-provided (#41)
* Switch from Python DevContainer feature to OS-provided

The DevContainer feature "python" always re-builds Python from source. This is a big issue for the build performance of the container, especially for cross-builds for arm64. Hence, we switch  to use the OS-provided Python. Note that this implies a change in patch version (from "3.12.11" to "3.12.3"), which should not be an issue.

* fix: install fixed git-lfs version

* minor cleanup

* fix: check if builder already exists