Try to fix multiarch builds of containers & speedup container builds (#43)
* bump version of docker-in-docker

* extend the publish script to multiarch

There are problems with creating multiarch devcontainers in one go via buildx and docker. One of the several solutions is to build and push per-architecture-tags, then create a merged metadata file with *all* architectures in one tag. Not perfect, but fits the current structure best (with docker-in-docker etc.)

* replace git and git-lfs features with apt calls

this dramatically speeds up the builds, and the exact tool versions probably do not matter that much.